+++ src/main/java/com/besson/arknights/block/custom/hr/OverloadedPrinter.java
@@ -1,0 +1,21 @@
+package com.besson.arknights.block.custom.hr;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.world.BlockView;
+
+public class OverloadedPrinter extends FurnitureHorizontalFacingBlock{
+    public static final VoxelShape SHAPE = Block.createCuboidShape(0,0,0,16,24,16);
+
+    public OverloadedPrinter(Settings settings) {
+        super(settings);
+    }
+
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return SHAPE;
+    }
+}